+++ hs-bindgen/test-artefacts/fixtures/macros/object_like_as_function_like.2.raw/Example.hs
@@ -4,14 +4,17 @@ module Example
     )
   where
 
+import qualified HsBindgen.Runtime.Macro as Macro
+import qualified HsBindgen.Runtime.Support as BG
+
 {-| __C declaration:__ @macro F@
 
     __defined at:__ @macros\/object_like_as_function_like.h 10:9@
 
     __exported by:__ @macros\/object_like_as_function_like.h@
 -}
-f :: [String]
-f = ["(", "x", ",", "y", ")", "x", "+", "y"]
+f :: Macro.Raw BG.Text
+f = Macro.functionLike "F" ["x", "y"] ["x", "+", "y"]
 
 {-| __C declaration:__ @macro G@
 
@@ -19,5 +22,6 @@ f = ["(", "x", ",", "y", ")", "x", "+", "y"]
 
     __exported by:__ @macros\/object_like_as_function_like.h@
 -}
-g :: [String]
-g = ["(", "x", ",", "y", ")", "x", "+", "y"]
+g :: Macro.Raw BG.Text
+g =
+  Macro.objectLike "G" ["(", "x", ",", "y", ")", "x", "+", "y"]